PT_containerization/TG_bot/requirements.txt
aiogram==2.17
python-dotenv
psycopg2-binary
paramiko

